requirements.txt
telethon>=1.28.3
python-dotenv>=1.0.0